Properly utilize LinkCache in SpecialAllPages
Viewing Special:AllPages currently results in one database query per
page displayed in the list. This is suboptimal because the page already
prefetches the required data for these pages, it just doesn't add it to
LinkCache, so LinkRenderer is unable to utilize it. Fix it by using
PageStore to fetch the list of pages to display, which handles this for
us. Do some minor cleanup in the affected area as appropriate.

Bug: T328174
Change-Id: I97b116e55997d78e93cbcf77a30e91f79f0545f7