specials: Change type documention to UserIdentity on Special:UserRights
Move more function to document the UserIdentity interface instead of
User or UserRightProxy.
In 1ecbfca calls were changed to need only UserIdentity type.
This makes it necessary to build the title for the userpage by the code
itself and not calling User::getUserPage or
UserRightsProxy::getUserPage. Both implementation have provided local
pages, because the value is stored in the local database.

Inline $isUserInstance to make phan happy

Change-Id: I3ddb4fd44d3d8c9f9a19fb5cfb3172fc42cb2327