MessageCache: Add a new hook for more performant override fetching
This new hook is called only once, for all overrides, instead of
every time a message is fetched.
Overrides can be specified as string or a callable, to allow for
dynamically computing the override on-demand. These callables are
passed the message key, MessageCache instance, the language and a
boolean indicating if the message should be fetched from the
database.

The parameter choices are based upon current usages of the
MessageCache::get hook.

Bug: T328183
Change-Id: I89e1a9b901aff6ff1bde729df8958ba42a45f3d5