vue i18n: Prevent unnecessary rerender in v-i18n-html
When the v-i18n-html directive is used in a component, the directive
function is called every time anything changes in the component.
Unfortunately this is necessary for detecting when the value of the
directive changes, because Vue's directive API doesn't give us a better
way to detect that.

However, it does give us the old value. This change adds logic that
should have been part of this directive from the start, comparing the
old value to the new value and doing nothing if they're the same. This
avoids unnecessary rerenders when an unrelated bit of component state
changes. These rerenders were wasteful, and were also causing bugs like
T327229.

There's no binding.oldArg, so we can't detect changes in binding.arg.
Those can only happen when a dynamic argument is used, which is already
discouraged, so warn in the documentation that that isn't supported.

Change-Id: I71f1cc0d08bfd02e1a9edb6cbfd849f10f929f3c