Fix XML dumps for content types with non-string getNativeData()
In fdc3e9f9524d91a492bdc212486d4518991c0fe2, the code generating XML
dumps was updated to support multi-content revisions. This refactor
included a workaround for content types that are subclasses of
TextContent to use getNativeData() rather than serialize(), apparently
to satisfy the Flow extension.

However, this assumes that getNativeData() always returns a string. As
demonstrated in T155582, this is not the case, which is one of the
reasons why the method was deprecated. Notably, if a wiki has a custom
content type defined whose getNativeData() returns a non-string value,
and has pages using that content type, this breaks XML dump generation
(dumpBackup.php) for that wiki and also makes those pages unexportable
via Special:Export.

Fix it by using getText() instead of getNativeData(), which is the
recommended migration path anyways per T155582. I am somewhat perplexed
by the reference to Flow in the original code comment, because Flow's
BoardContent does not seem to extend TextContent at all.

Bug: T155582
Bug: T328503
Change-Id: I670fb53f193ec20d3d4c258e54c89e7f64cf2d1b