Add more items to MediaWiki:metadata-fields
This controls which metadata fields are shown by default on
the image description page. The intention is to only show fields
likely to be of interest to users, and hide the more technical
fields. This list was created very early in MediaWiki history, and
there are a lot more fields that are probably of interest. Some
of these fields are rather obscure and not used on very many
files. The field will only be shown if it is in use.

Add the fields that either describe the content of the image, or
describe the copyright status of the image.

Change-Id: Iefc4dddee2a42464eccf6d300e361bcc26f92574