SpecialUserrights: Replace User instance checks with UserIdentity::LOCAL
Do not assume that when a User object is given,
this is a local user right change. Check the wiki id instead and
call the UserFactory to create a local User object, if needed.
In the long term this allows to replace UserRightsProxy functions by a
wiki-aware UserIdentityLookup (via ActorStore)

Change-Id: Ibd23b36680bc0c87e27c6c2518dccc36b7f73518