Do not use media filename as alt attribute
Matches Parsoid's current output.

Not a canonical source, but, this site says,
https://help.siteimprove.com/support/solutions/articles/80000863904-accessibility-image-alt-text-best-practices

> If no alt attribute is present, the screen reader will read the file
> name for the image instead, which can be a major distraction to those
> using screen-reading technology.

So, reading the filename seems to be a default behaviour anyways and
using the filename doesn't seem to adding any benefit.  However, placing
it preempts any improvements that might happen in screen reading
technology since the screen reader would likely prefer the alt attribute
to any magic it tries to do in its absence (like machine vision
processing of the image).

An alternative proposal would be to strip off the file extension as in
I218e5565816b7643f3b85083031644e3e4749a5c and implement the same in
Parsoid.

Longer term plans that actually address the issue here are in T325955
and T63566.

Bug: T326041
Bug: T63566
Bug: T325955
Depends-On: I7b1f07190e8eaca5cbda38d9ce366aa60041ab81
Depends-On: I9dd37f70be8163df76c154f175ef50134fb811d8
Depends-On: If9cdabdfac26656272fcf3b4aaae0576aaed1346
Change-Id: If1e55feb86ce8b32f772e3b78bc9d29f122f4d58