phpunit: Add PHPUNIT_LOGS flag
Dumping logs can be useful for interpreting CI test failures, but
is rarely needed when running phpunit locally and it makes the
output unreadable. In theory it should be possible to disable them
with the --no-extensions flag but that doesn't seem to work.

Allow disabling log dumping by prefixing the phpunit command with
PHPUNIT_LOGS=0.

Change-Id: I071ff3228e922c9aaf705e32e9e1a2351bb1e973