LinkFilter: Get rid of concept of field prefix in ::getQueryConditions
This made some sense in tables like pagelinks or templatelinks where
there are several tables with similar schemas (it's better to use a
central class intead but understandable).

For externallinks, it's completely useless. There is no other table with
_index_60 field and as result, no code actual use this option either.

It makes grepping harder, it's prone to errors and complicates the code
more than it needs to be.

Bug: T326251
Change-Id: I72ad81d5c3e8f62ceee30d9a44d71f37655ef506