SpecialUserrights: Read groups from primary DB during conflict check
To have a working conflict check, the UserGroupManager should read from
the primary. There is already a check that the request was posted before
this check is reached, so it is save to query the primary.

For local user rights, the call to User::clearInstanceCache has cleared
some caches in UserGroupManager, which makes the manager reload the
group from (any) database. But using the "latest" option is much easier
and also works for cross-wiki userright changes.

Change-Id: Ie290def7499b51932f183c4472193e7e9e886b11