ParserTestRunner: Move 'showflags' handling inside ::addParserOutputInfo()
This is just a clean up refactor to keep the option handling code in one
place, and to allow Parsoid to share it when running in integrated mode.

But in the process we tweaked a few other things for consistency:

* The 'pst' option now can add all of the parser output metadata flags
  (not just 'showflags')

* The 'ill' and 'cat' options append to, instead of replacing, the
  parser output.

* A new 'nohtml' option was added to allow suppressing the HTML in
  cases where replacing, rather than appending to, the HTML is
  desired.

* 'showflags' doesn't emit an 'extra' newline at the start of its output
  any more.

The appropriate changes to parser tests were performed to accomodate
these tweaks.

Code search for uses of these options (uses in parsoid and core can be
ignored; uses of 'pst' are largely unchanged):

  https://codesearch.wmcloud.org/search/?q=%28%5E%7C%20%29%28pst%7Cill%7Ccat%7Cshowflags%29%28%20%7C%24%29&i=nope&files=%5Etests%2Fparser%2F.*%5C.txt&excludeFiles=&repos=

Depends-On: I5c22f456a3ae5ea25b59c4246d68965099c465cc
Depends-On: If307de5d683829beb552663c72288d065827cfb6
Change-Id: I61a67a0e6928463e3872be9a42ff6992c6754662