ApiResult: Make array ordering consistent across PHP versions
PHP 8.2 makes krsort() use the stricter numeric comparison semantics[1]
introduced in PHP 8.0 when it is comparing a numeric key with a
non-numeric key. This changes result ordering in ApiResult for arrays
that contain a mix of numeric and non-numeric keys.

Make the ordering consistent across PHP versions by using an
user-defined comparison function instead that always considers a
non-numeric string to be smaller or equal in a comparison of a
non-numeric and a numeric string.

[1] https://wiki.php.net/rfc/string_to_number_comparison

Bug: T326480
Change-Id: I91d55a1d516e4f0ec31f00a719c6616c73b7f786
(cherry picked from commit 538fb85a17e2189fcd5792edb1aa1893a03805b5)