Tests: Error for any new module that introduces bad targets
Going forward we do not want to use the targets system in the way
we traditionally used it. While updating existing usages will take
time it seems like it might be a good idea to prevent additional usages
being added. This list would gradually be updated as we update those
components.

Bug: T127268
Depends-On: Ibd7270abc5f223bc4f4814c896e7ef4d02d21386
Change-Id: I2e226c174be88caeb505738645cfa0582e39a197