Make use of jQuery method chaining where it makes sense
It's not always better to arrange code this way because it makes the
lines depend on each other and makes it a little harder to change one
line without changing the others. But I feel like it makes a lot of
sense in these cases. These method calls are not independend from
each other.

Change-Id: I6b223018d900b907933e9d4ed010af47f92ed37d