Sync up core repo with Parsoid
One of the toc.txt tests imported from Parsoid had a dependency on the
Cite extension, which is not necessarily installed when core runs its
parser tests.  Disabled that test for now; perhaps a per-test version
of the `!!hooks` clause can be added to handle situations like this
in the future.

This now aligns with Parsoid commit b4a2b75c2d4384a9d1c1eaa553332b566cd2d183

Change-Id: Ifd4f258742d83c879f724b9a61cb309c999c6d65