Remove obsolete Rdbms query groups from getConnection calls
The currently supported groups are documented in database.md.

While at it, prefer the string form of the $groups parameter where
possible, since the complexity of an array is never utilized, as the
groups either can't overlap, or are decided on by the caller instead.

This also helps with the simpler LBFactory::getConnection() proposed
in T326274.

Bug: T263127
Bug: T326274
Change-Id: I7a31a07a8c3f1e265ac43ab59f23326590b407c3