title: Use Title::makeName instead of ::getFullText in ::isValid
In getFulltext the namespace number is converted to text, for NS_USER
this trigger the gender cache. Use canonical namespace instead.

When Title::canExist is called to fill a LinkBatch (Like done for
Special:Log/block via BlockLogFormatter::getPreloadTitles),
the effect of the batch lookup is not the best, because GenderCache
filled with single database queries instead.

This is similar to the usage of makeName in Title::makeTitleSafe

Handle invalid namespace numbers in Title::makeName the same as in
Title::prefix

Change-Id: Ibbdc7e0762568bb29ae59764c3666f275d201c6c