rdbms: simplify query group selection in LoadBalancer::getConnection()
Make resolveGroups() pick only the first query group with defined
servers. Remove the fallback logic if all of those servers are down.

This simplifies the code, avoids the expense of trying to reconnect on
ever call to getConnection(), and reduces the chance of queryies suddenly
being routed to another server in the middle of a request (e.g. when the
server for group becomes reacheable again). Switching servers means that
queries no longer use the same point-in-time REPEATABLE-READ snapshot.

Bug: T326187
Change-Id: I2907743163d84a6665902d2591e1eab2b492f96a