WikiPage: Simplify added/deleted logic in ::updateCategoryCounts()
Follows-up Ic3f319e0a0f.

The $added and $deleted params are known to always be arrays by
strict type hint, so there is no need to call count() to know
whether they are non-empty.

Improve comprehension of intersect/diff by bringing the statements
together.

Fix the $removeFields query to use the same variable for its condition
as for its query as there is no need to run one if the other is empty
instead. (In practice, that's not possible, but it removes doubt
from the code and makes it more obviously correct.)

Bug: T300303
Change-Id: I4f79d2a152068290b732753030ad2705a201138f