EnhancedChangesList: Isolate categorization entries in getLogText
rc_this_oldid and rc_last_oldid of categorization entries
reference revisions of the page that had its categorization
changed rather than the category page itself. These should
not be considered when building the diff link(s).

Introduce new aggregate variable $allCategorization and use
it instead of examining $rcObj (which deals only with the last
change in the block).

Finally, drop the ::userCan check deciding whether the user
can see the diff. It wouldn't work properly either, because
it would check only the last (oldest) entry in the block, not
the first (newest) and the one before the last.

Bug: T261068
Change-Id: I10bdfbad9d8a36f210ffd11a832604104f8ecdc5