Do not replace all whitespace characters with space in tablesorter
Many languages use non-breaking space as the thousands separator, and
replacing the nbsp with a normal space causes the number to not be
recognised as such.

The regression was introduced in
I013a05a4a4a2fc16f74d1c858f6706e986d96a4e.

This patch is meant to be a quick fix to the issue without having to
revert all the involved patches, but I'm still not entirely confident
about this code.

Bug: T328218
Change-Id: I62eb7f8b43788b78d24dcb53cec3ec6fb87354a7