Add --before-timestamp option to refreshLinks.php
We want to be able to refresh pages that haven't been updated in quite a
while to ensure any MediaWiki parsing changes, etc. get reflected in
links tables. This information is already tracked in the
page.page_links_updated field, which we'll filter with. We can't
actually do it in SQL because there's no index on the column, but it
gets loaded by WikiPageFactory::newFromID(), so we simply check it
in the fixRedirect() and fixLinksFromArticle() functions.

== Test plan ==
* Run plain `refreshLinks.php` and see that all page_links_updated
  fields have been updated to now (this patch doesn't break existing
  functionality).
* Backdate some page_links_updated fields.
* Run refreshLinks.php --before-timestamp X, where X is between your
  backdated values and now.
* Observe that only the backdated pages have had their
  page_links_updated modified to now.

Bug: T159512
Change-Id: I695d971ef7cbabddda3125361975be0f94dabf4c