Template list in previews: Look up missing restriction level messages
When the template list is shown on preview, the protection level is
looked up from the API and so the i18n messages for these may not
already be available in the front end. This adds an extra API call
to fetch any missing messages, and refactors the other code to
make this possible.

The templateinfo call is also switched to a POST request in order
to avoid 414 Request-URI Too Long errors when there are lots of
template titles passed.

Bug: T321032
Change-Id: I5f0ec980c64d7ea025cbca31eeb5e4ea1b4bc360