changeslist: Avoid rendering the page if performing an HTTP redirect
If an user has an active RCFilter, RecentChanges and Watchlist redirect
them from the base page URL (e.g. Special:RecentChanges) to the full URL
including the query params for that active filter (e.g.
Special:RecentChanges?urlversion=2&hidebots=1&...). However, the full
page is still generated, even though the response is just an HTTP
redirect. This is suboptimal, as the DB queries and HTML rendering
needed to render these pages is expensive.

As a fix, short-circuit execution if the page is going to perform an
HTTP redirect to avoid spending time rendering an UI that will not be
shown.

Bug: T330100
Change-Id: Ia9b7d5dad75ca4ddc43df90b26385dc801435ed5