Suppress false SecurityCheckXSS positives in ParserTestPrinter
The parser test runner is a CLI tool; all the XSS warnings are
false positives.  There were a couple of individual suppressions
already in the file, but applied the advice from the upstream
SecurityCheckPlugin repo of routing all potentially tainted
output through a dedicated `::print` function so that the
taintedness can be suppressed in a single place.

Change-Id: I8023931d31e494df325b16d1b922539e20b58c51