mediawiki.storage: Disable async GC during integration test
Tests may trigger GC explicitly, but avoid doing so in the background
made tests non-deterministic in terms of what they observe in
localStorage. It also meant that setTimeout, mw.storage,
mw.requestIdleCalback, $.Deferred() etc would be called during a
random unrelated test and thus invoke stubs and mocks that either
make the GC fail in a CI-breaking way, or make the caller break
due to untimely invoking a mock or stub.

Bug: T318103
Change-Id: I5fbc0ee2fdb55b631bce57022cc15d49fa85b110