build: Update eslint-plugin-qunit to ^7.3.0
The 7.3.0 version, released in March 2022, fixed an issue where the
`hooks` parameter to `QUnit.module` sometimes received a unexpected
lint warning due to misunderstanding what the parameter is used for.

https://github.com/platinumazure/eslint-plugin-qunit/pull/231

The existing range ^7.2.0 range, as dependency of
eslint-config-wikimedia, would already expand to 7.3.4 if we build
package-lock.json, but since npm is conservative nowadays, it hasn't
been bumped yet.

Several other MediaWiki repos already use the newer version, which
likely did a rebuild at some point in the last 11 months.

<https://codesearch.wmcloud.org/deployed/?q=eslint-plugin-qunit-7.3...tg>

Change-Id: I696ed00372f29b4e1a9c67e71e5af03c9c7acd54