qunit: Remove old $.fx and $.ajax tracking from testrunner
I added this about ten years ago after I encountered two issues in
the same week where, during development, I couldn't get a newly written
test to pass. The root cause was that the test was async, but I
forgot to use `assert.async()`. In one case due to an XHR request,
and in one case due to an unfinished animation. By tracking these,
and warning about unfinished ones, this would be found sooner.

To my knowledge, nobody (including myself) has ever since then
encountered these warnings during development in a way that helped
identify a mistake. Instead, we've only seen them ocasionally pop
up in CI due to unfinished AJAX requests and animations unrelated
to tests.

It is best practice not to trigger animations or requests during a test
that you don't wait for. In practice, I believe such a thing would
likely be a mistake and only make the test slower. But, we can
follow that best practice without enforcing it, with the benefit
of tolerating a few minor inefficiencies here and there, without
causing CI to be in a failing state every time it slips in, especially
if it slips in in a non-deterministic way, causing flaky tests.

But.. I didn't introduce it to help with that best practice, and
it seems we've been pretty good about this regardless. I introduced
it to help debug a test that is consistently failing on code that
isn't merged yet. In retrospect, I think a modern DevTools/debugger
would help you figure out the issue easy enough with a breakpoint
without this additional complexity.

The objective here is to make the QUnit test runner simpler and less
special in ways that are specific to MediaWiki, as well as to make it
a bit faster by reducing logic that has to run for every single test
case.

Bug: T250045
Change-Id: I77878f2d855da89f6b2e2c0e17060daacf867678