Use same formula to construct cache key in increase() and clear()
The cache key formula was updated in I7e19cd0d9 to replace null $ipKey
and $userKey values with empty strings in increase(). However, clear()
was not similarly updated in that patch. This patch updates both
functions to share the code from the earlier patch. Due to that change,
this patch also fixes a deprecation warning emitted in BagOStuff when
it calls strtr when using PHP8 when $ipKey or $userKey is null in
clear().

Change-Id: Ia597c9523f78a638fccae37ac0a61fc20e128716