EnhancedChangesList: Rewrite bogus condition
It was definitely already wrong when it was added in 2008
(r33013). !$type will return true for RC_EDIT (!0), false for
all other types (RC_NEW, RC_LOG, etc.). The weak equality
comparison then casts RC_LOG to boolean (3 -> true).
So the first condition will pass if and only if $type is
RC_EDIT.

As this piece of code lasted for 15 years and there
seems to be no objection to the behavior, I just
rewrite it to make it more clear and keep the same
behavior.

Cudos to Thiemo Kreuz who disassembled the code.

Change-Id: I93f853b2a191fc3011b47fcd1ac22ac649ce541a