Follow-up fixes for SpecialRevisionDelete OOUI conversion
Follow-up fixes / changes for the conversion of
Special:RevisionDelete to OOUI. This patch includes:
* Changing "(do not change)" to "Do not change"
* Only providing "options" in the field defintion if the type is
a radio and not in the case of a checkbox
* Adding two new messages that make the radio options to suppress
clearer (specifically using "Suppress" and "Unsuppress" instead of
"Hidden" and "Visible").
* Update qqq.json to reflect the changes and to use a new image
that shows the OOUI interface.

None of these issues prevented merging of the patch this follows up
and to avoid another merge conflict I decided merging was okay.

Bug: T117794
Follow-Up: I802a064572395a7373bc726fb477b5be2ee169ce
Change-Id: I7071ba81e92f26443536544f58b74bbb74c338fc