mediawiki.Uri: Document potential for corrupting the fragment
mw.Uri uses encodeURIComponent() for encoding the fragment, which
percent-encodes more characters than it should, including the /
character which is often used in routes the fragment is matched
agains. This means that URL manipulation using mw.Uri with and
something like history.replace might break routing.

Mention this in the documentation of mw.Uri.toString() and refer
to URL which handles fragment roundtripping according to the
WHATWG URL spec.

Change-Id: Ib17349308359665b1ed2c28b91e15207f6fa966c