Remove MainConfigNames.template.
generateConfigSchema.php was already generating all other files
directly, MainConfigNames.php was the only one based on the template.
Given the template was trivial and unlikely to change, this change
removes the template and changes the script to generate
MainConfigNames.php directly as well, in the name of consistency and
simplicity.

Change-Id: I7e7f2bfbc232d2ee6582638ced74b3e929be21ed