mediawiki.skin: Provide wrapper files for importing Codex design tokens
Skins can't reliably import the Codex tokens for their theme directly,
because the relative path from the skin's directory to the resources/lib
directory isn't the same in every setup. Work around this issue by
providing wrapper files in a subdirectory of mediawiki.less, so that
skins can import the Codex design tokens with
@import 'mediawiki.skin.codex-design-tokens/theme-themename.less';

Unfortunately this requires providing a file for each theme. I tried
to avoid this by writing a mixin that dynamically imports the right
theme file given a theme name, but it turns out imports in mixins are
applied after top-level imports, so the fallback values in the defaults
file ended up overriding the values from the theme file rather than the
other way around.

Bug: T325237
Change-Id: I3a8c89d8558022077be1f32c6a6f4733b6302948